android_filesystem_config: Fix compiles on master branch
pointer type carried over by mistake as I was editing on the native branch originally (added in https://github.com/osm0sis/mkbootfs/commit/2b28b7863199a523e5e7981892970d27a7471d34)

+++ private/android_filesystem_config.h
@@ -349,7 +349,7 @@ static const struct fs_path_config android_files[] = {
 };
 
 static inline void fs_config(const char *path, int dir,
-                             long unsigned *uid, long unsigned *gid, unsigned *mode, uint64_t *capabilities)
+                             unsigned *uid, unsigned *gid, unsigned *mode, uint64_t *capabilities)
 {
     const struct fs_path_config *pc;
     int plen;
